feat: refactor tiles
