+++ src/components/Grid/grid.scss
@@ -9,14 +9,24 @@
     padding-bottom: rem(80px);
     padding-top: rem(56px);
   }
+
+  &:last-child {
+    padding-bottom: rem(128px);
+  }
+}
+
+.background--nomargin {
+  padding-top: rem(32px);
+  padding-bottom: 0;
+
+  &:last-child {
+    padding-bottom: 0;
+    margin-bottom: rem(128px);
+  }
 }
 
 .page-content > div div.background:first-child {
   padding-top: 2rem;
-}
-
-.background:last-child {
-  padding-bottom: rem(128px);
 }
 
 .background--white {
@@ -33,17 +43,17 @@
 
 .background--gray-80 {
   background: $ibm-colors__gray--80;
-  color: $ibm-colors__white;
+  color: $ibm-colors__gray--10;
 }
 
 .background--gray-100 {
   background: $ibm-colors__gray--100;
-  color: $ibm-colors__white;
+  color: $ibm-colors__gray--10;
 }
 
 .background--black {
   background: $ibm-colors__black;
-  color: $ibm-colors__white;
+  color: $ibm-colors__gray--10;
 }
 
 .background--gray-white {
