+++ src/content/approach/approach.md
@@ -32,37 +32,60 @@ responsibility to the people we serve.</p>
 </column>
 
 </grid>
+<grid background="black" margin="false">
+<column lg="4" offset_lg="1">
+
+<h2><strong>Design Philosophy</strong><br>The beliefs behind everything we do.</h2>
+
+Design is about giving people a path, both emotionally and functionally, towards their goals.
+
+</column>
+<column lg="8" offset_lg="3">
 
 <tile
-    size="xl"
-    background="#121212"
-    title="Design Philosophy"
-    title_secondary="The beliefs behind everything we do."
-    description="Design is about giving people a path, both emotionally and functionally, towards their goals."
-    icon="ArrowRight16"
-    href="/approach/design-philosophy">
+    href="/approach/design-philosophy"
+    caption="caption"
+    title="title">
 <img src="images/Image_1.svg" alt="Geometric shapes"/>
 </tile>
 
+</column>
+</grid>
+<grid background="gray-20" margin="false">
+<column lg="4" offset_lg="1">
+
+<h2><strong>Design Thinking</strong><br>Human-centered design at scale.</h2>
+
+Learn how you can apply the same framework our teams use every day.
+
+</column>
+<column lg="8" offset_lg="3">
+
 <tile
-    size="xl"
-    background="#E7E7E7"
-    title_color="dark"
-    title="Design Thinking"
-    title_secondary="Human-centered design at scale."
-    description="Learn how you can apply the same framework our teams use every day."
-    icon="ArrowRight16"
-    href="/approach/design-thinking">
+    href="/approach/design-thinking"
+    caption="caption"
+    title="title">
 <img src="images/Image_2.svg" alt="Geometric shapes"/>
 </tile>
 
+</column>
+</grid>
+<grid background="gray-80" margin="false">
+<column lg="4" offset_lg="1">
+
+<h2><strong>Design Services</strong><br>Your business partner by design.</h2>
+
+Partner with us to help define your strategy, create exceptional experiences, and drive better business outcomes.
+
+</column>
+<column lg="8" offset_lg="3">
+
 <tile
-    size="xl"
-    background="#373737"
-    title="Design Services"
-    title_secondary="Your business partner by design."
-    description="Partner with us to help define your strategy, create exceptional experiences, and drive better business outcomes."
-    icon="ArrowRight16"
-    href="/approach/design-services">
+    href="/approach/design-services"
+    caption="caption"
+    title="title">
 <img src="images/Image_3.svg" alt="Ven diagram emphasizing overlapping area"/>
 </tile>
+
+</column>
+</grid>
