+++ src/content/approach/design-thinking/design-thinking.md
@@ -136,16 +136,19 @@ Sponsor users are real-world users that provide teams with deep expertise and kn
 
 </column>
 </grid>
+<grid background="gray-10">
+<column lg="16">
 
 <tile
-  size="md"
-  background="#C5DEFF"
-  title="Explore how you and your business can take advantage of our entire platform."
-  name="Enterprise Design Thinking">
+    href="#"
+    title="Enterprise Design Thinking"
+    feature="true"
+    feature_heading="Explore how you and your business can take advantage of our entire platform."
+    feature_background="black">
 <img src="images/Image_2.png" alt="Geometric shapes"/>
 </tile>
 
-<grid background="gray-10">
+</column>
 <column lg="8">
 
 <h3>Keep exploring<br>our approach</h3>
