+++ src/content/demo/demo.md
@@ -2,146 +2,133 @@
 title: Demo
 ---
 
-<grid background="gray-10">
+<grid background="black" margin="false">
 <column lg="4">
 
-## Small tile w/ image
+## **Design Philosophy** the beliefs behind everything we do
 
-<tile
-    caption="caption"
-    name="name"
-    href="#">
-<img src="../images/Image_1.png" alt="A laptop showing security analytics" class="tile__img"/>
-</tile>
+Design is about giving people a path,
+both emotionally and functionally,
+towards their goals.
 
 </column>
-<column lg="4">
-
-## Small dark
+<column lg="12">
 
 <tile
+    href="#"
     caption="caption"
-    name="name"
-    dark="true"
-    href="#">
-<img src="../images/Image_1.png" alt="A laptop showing security analytics" class="tile__img"/>
-</tile>
-
-</column>
-<column lg="4">
-
-## Small tile w/ out an image
-
-<tile
-    caption="caption"
-    name="name"
-    href="#">
+    title="title">
+<img src="../approach/images/Image_1.svg" alt="Geometric shapes"/>
 </tile>
 
 </column>
 </grid>
-<grid background="white">
+<grid background="white" margin="false">
+<column lg="4">
+
+## **Design Philosophy** the beliefs behind everything we do
+
+Design is about giving people a path,
+both emotionally and functionally,
+towards their goals.
+
+</column>
+<column lg="12">
+
+<tile
+    href="#"
+    caption="caption"
+    title="title"
+    dark="true">
+<img src="../approach/images/Image_1.svg" alt="Geometric shapes"/>
+</tile>
+
+</column>
+</grid>
+<grid background="gray-10">
 <column lg="16">
 
-## Md Tile
+<tile
+    href="#"
+    caption="caption on feature"
+    title="title on feature"
+    feature="true"
+    feature_heading="Feature Heading"
+    feature_background="black">
+<img src="Image_2.png" alt="Geometric shapes"/>
+</tile>
 
 </column>
 </grid>
-
-<tile
-    size="md"
-    background="gray-100"
-    light="true"
-    title="title"
-    name="name"
-    caption="caption"
-    title_secondary="title_secondary"
-    description="description"
-    icon="ArrowRight16"
-    href="/approach/design-philosophy">
-<img src="../approach/design-services/images/Image_2.png" alt="alt"/>
-</tile>
-
-<grid background="white">
+<grid background="gray-10">
 <column lg="16">
 
-## Md Dark
+## small tiles
 
 </column>
-</grid>
-
-<tile
-    size="md"
-    background="white"
-    dark="true"
-    title="title"
-    name="name"
-    caption="caption"
-    title_secondary="title_secondary"
-    description="description"
-    icon="ArrowRight16"
-    href="/approach/design-philosophy">
-<img src="../approach/design-services/images/Image_2.png" alt="alt"/>
-</tile>
-
-<grid background="gray-10">
 <column lg="12" offset_lg="4">
 
-## Large Tile
+<tile
+    href="#"
+    caption="caption"
+    title="tile with image">
+<img src="../practices/images/Image_5.png" alt="Geometric shapes"/>
+</tile>
+
+</column>
+<column lg="16">
 
 <tile
-    size="lg"
-    background="#C5DEFF"
-    light="true"
+    href="#"
     caption="caption"
-    name="name">
+    title="tile with image">
+<img src="../practices/images/Image_5.png" alt="Geometric shapes"/>
+</tile>
+
+</column>
+
+<column lg="4" sm="2">
+
+<tile
+    caption="caption"
+    title="title image"
+    href="#">
+<img src="../images/Image_1.png" alt="A laptop showing security analytics"/>
+</tile>
+
+</column>
+<column lg="4" sm="2">
+
+<tile
+    caption="caption"
+    title="title"
+    dark="true"
+    href="#">
+<img src="../images/Image_1.png" alt="A laptop showing security analytics"/>
+</tile>
+
+</column>
+<column lg="4" sm="2">
+
+<tile
+    caption="caption"
+    title="title"
+    href="#">
+</tile>
+
+</column>
+
+</grid>
+
+<grid background="white">
+<column lg="16">
+
+<tile
+    href="#"
+    caption="caption"
+    title="title">
 <img src="../practices/images/Image_5.png" alt="Geometric shapes"/>
 </tile>
 
 </column>
 </grid>
-
-<grid background="gray-10">
-<column lg="12" offset_lg="4">
-
-## Large dark
-
-<tile
-    size="lg"
-    background="#C5DEFF"
-    dark="true"
-    caption="caption"
-    name="name">
-<img src="../practices/images/Image_5.png" alt="Geometric shapes"/>
-</tile>
-
-</column>
-</grid>
-
-## XL Tile
-
-<tile
-    size="xl"
-    background="gray-100"
-    light="true"
-    title="title"
-    title_secondary="title_secondary"
-    description="description"
-    icon="ArrowRight16"
-    href="/approach/design-philosophy">
-<img src="../approach/images/Image_1.svg" alt="Geometric shapes"/>
-</tile>
-
-## XL dark
-
-<tile
-    size="xl"
-    background="white"
-    dark="true"
-    title="title"
-    title_secondary="title_secondary"
-    description="description"
-    icon="ArrowRight16"
-    href="/approach/design-philosophy">
-<img src="../approach/images/Image_1.svg" alt="Geometric shapes"/>
-</tile>
