+++ src/content/practices/practices.md
@@ -24,7 +24,7 @@ building bonds.</span>
 
 <p size="xl">Tap into a growing collection of design tools, assets and resources to keep your business moving forward.</p>
 
-<icon name="ArrowDown32" color="white"></icon>
+<icon title="ArrowDown32" color="white"></icon>
 
 </column>
 </grid>
@@ -44,40 +44,14 @@ building bonds.</span>
 ## Principles, guidelines and ready-made assets to help you build exceptional experiences.
 
 </column>
-<column lg="2" md="2" offset_lg="4">
 
-<tile
-    dark="true"
-    size="xs"
-    name="Gallery">
-</tile>
-
-</column>
-<column lg="2" md="2">
-
-<tile
-    dark="true"
-    size="xs"
-    name="Elements">
-</tile>
-
-</column>
-<column lg="2" md="2">
-
-<tile
-    dark="true"
-    size="xs"
-    name="System">
-</tile>
-
-</column>
 <column lg="12" offset_lg="4">
 
 <tile
     size="lg"
     background="#373737"
     dark="true"
-    name="IBM Design Language">
+    title="IBM Design Language">
 <img src="images/Image_1.png" alt="Geometric shapes"/>
 </tile>
 
@@ -86,7 +60,7 @@ building bonds.</span>
 
 <tile
     dark="true"
-    name="Product Design System">
+    title="Product Design System">
 <img src="images/Image_2.png" alt=""/>
 </tile>
 
@@ -95,7 +69,7 @@ building bonds.</span>
 
 <tile
     dark="true"
-    name="Digital Design System">
+    title="Digital Design System">
 <img src="images/Image_3.png" alt=""/>
 </tile>
 
@@ -104,7 +78,7 @@ building bonds.</span>
 
 <tile
     dark="true"
-    name="IBM Plex typeface">
+    title="IBM Plex typeface">
 <img src="images/Image_4.png" alt=""/>
 </tile>
 
@@ -126,40 +100,13 @@ building bonds.</span>
 ## A scalable learning approach that combines online learning with in-person activation to make design thinking stick.
 
 </column>
-<column lg="2" md="2" offset_lg="4">
-
-<tile
-    dark="true"
-    size="xs"
-    name="Badges">
-</tile>
-
-</column>
-<column lg="2" md="2">
-
-<tile
-    dark="true"
-    size="xs"
-    name="Toolkit">
-</tile>
-
-</column>
-<column lg="2" md="2">
-
-<tile
-    dark="true"
-    size="xs"
-    name="FAQ">
-</tile>
-
-</column>
 <column lg="12" offset_lg="4">
 
 <tile
     size="lg"
     background="#C5DEFF"
     dark="true"
-    name="Enterprise Design Thinking">
+    title="Enterprise Design Thinking">
 <img src="images/Image_5.png" alt="Geometric shapes"/>
 </tile>
 
@@ -181,40 +128,13 @@ building bonds.</span>
 ## Resources and guidance to help you drive actionable user insights through observation, experience, and making.
 
 </column>
-<column lg="2" md="2" offset_lg="4">
-
-<tile
-    dark="true"
-    size="xs"
-    name="Principles">
-</tile>
-
-</column>
-<column lg="2" md="2">
-
-<tile
-    dark="true"
-    size="xs"
-    name="Foundation">
-</tile>
-
-</column>
-<column lg="2" md="2" md="2">
-
-<tile
-    dark="true"
-    size="xs"
-    name="Resources">
-</tile>
-
-</column>
 <column lg="12" offset_lg="4">
 
 <tile
     size="lg"
     background="#F2BDD8"
     dark="true"
-    name="IBM Design Research">
+    title="IBM Design Research">
 <img src="images/Image_6.png" alt="Geometric shapes"/>
 </tile>
