+++ src/styles/_page.scss
@@ -18,7 +18,7 @@ body {
 }
 
 .background--inverse a.bx--link:active {
-  color: $ibm-colors__white;
+  color: $ibm-colors__gray--10;
 
   svg {
     fill: $ibm-colors__white;
